Refactor code into MVC modules
