+++ index.html
@@ -71,6 +71,6 @@
 			</div>
 		</div>
 	</template>
-	<script src="index.js"></script>
+	<script src="index.js" type="module"></script>
 </body>
 </html>
